Switch to forge releases for test fixtures
Using forge modules instead of repositories with a certain ref makes
sure the unit tests are run against the latest and greatest version of
the dependency that is available at the forge. It is still possible to
restrict to a certain release version though if needed.
The symlink for the module under test itself is left in the fixture file
though not strictly needed as it is symlinked by default.